requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
mysqlclient==2.2.4
setuptools==72.1.0
sqlparse==0.5.1
wheel==0.43.0
